ensure we keep the server as busy as we can
ServerAction/Do_ToDo was flawed because it would not re-call Push_AllReqs
until there no jobs running; and so in some scenarious we would end up
with just one job running for quite a while (e.g. with lots of tests
to run and some of those tests running for a long time).

ServerAction/Do_ToDo was also unnecessarily convoluted.

So simplified it and in the process ensured we always launch jobs as
soon as they are ready and server has room.