python xju.json_codec allow non-str key types
... key type can be str or some union of int,float,bool,None,
xju.newtype.{Int,Float,Bool}

This follows python's interesting handling of dict keys:
it allows int/bool/float/None and encodes them as str. Interesting
because it then must be asymmetric (it cannot tell what to decode as
so simply leaves as str).

Note that the differentiation of str is because in json itself
all keys are strings, so a key type of str|int would be ambiguous.